refactor: entity name resolution service
