+++ waltid-wallet-api/src/main/kotlin/id/walt/webwallet/service/entity/DefaultNameResolutionService.kt
@@ -1,6 +1,7 @@
-package id.walt.webwallet.service.issuers
+package id.walt.webwallet.service.entity
 
 import id.walt.webwallet.config.TrustConfig
+import id.walt.webwallet.utils.JsonUtils
 import io.ktor.client.*
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
@@ -8,10 +9,10 @@ import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.jsonPrimitive
 
-class DefaultIssuerNameResolutionService(
+class DefaultNameResolutionService(
     private val http: HttpClient,
     trustRecord: TrustConfig.TrustRecord?,
-) : IssuerNameResolutionService {
+) : EntityNameResolutionService {
     private val baseUrl = trustRecord?.baseUrl ?: ""
     private val governanceRecordPath = trustRecord?.governanceRecordPath ?: ""
     private val json = Json {
@@ -21,8 +22,8 @@ class DefaultIssuerNameResolutionService(
     override suspend fun resolve(did: String): Result<String> = runCatching {
         http.get(String.format("$baseUrl/$governanceRecordPath", did)).bodyAsText().let {
             json.decodeFromString<JsonObject>(it).let {
-                it["name"]?.jsonPrimitive?.content
+                JsonUtils.tryGetData(it, "name")?.jsonPrimitive?.content
             }
-        } ?: error("Error resolving issuer name for: $did")
+        } ?: error("Error resolving entity name for: $did")
     }
 }
