+++ waltid-wallet-api/src/main/kotlin/id/walt/webwallet/service/entity/EntityNameResolutionService.kt
@@ -1,0 +1,5 @@
+package id.walt.webwallet.service.entity
+
+interface EntityNameResolutionService {
+    suspend fun resolve(did: String): Result<String>
+}
