--- waltid-wallet-api/src/main/kotlin/id/walt/webwallet/service/issuers/IssuerNameResolutionService.kt
@@ -1,5 +1,0 @@
-package id.walt.webwallet.service.issuers
-
-interface IssuerNameResolutionService {
-    suspend fun resolve(did: String): Result<String>
-}
